Move custom front page button to top-level
