+++ packages/lesswrong/components/themes/ThemePickerMenu.tsx
@@ -1,17 +1,11 @@
 import React from 'react';
-import { useUpdateCurrentUser } from '../hooks/useUpdateCurrentUser';
 import { ThemeMetadata, getThemeMetadata, AbstractThemeOptions } from '../../themes/themeNames';
-import { isEAForum, isLW, isLWorAF } from '../../lib/instanceSettings';
 import { ThemeContext } from './useTheme';
-import { HomeDesignChatContext } from '../common/HomeDesignChatContext';
-import { useCurrentUser } from '../common/withUser';
 import { isMobile } from '../../lib/utils/isMobile'
-import { useNavigate } from '@/lib/routeUtil';
 import { Paper }from '@/components/widgets/Paper';
 import LWTooltip from "../common/LWTooltip";
 import DropdownMenu from "../dropdowns/DropdownMenu";
 import DropdownItem from "../dropdowns/DropdownItem";
-import DropdownDivider from "../dropdowns/DropdownDivider";
 import ForumIcon from "../common/ForumIcon";
 import { defineStyles } from '@/components/hooks/defineStyles';
 import { useStyles } from '@/components/hooks/useStyles';
@@ -32,20 +26,8 @@ const ThemePickerMenu = ({children}: {
 }) => {
   const classes = useStyles(styles);
   const themeContext = React.useContext(ThemeContext)!;
-  const designChat = React.useContext(HomeDesignChatContext);
   const currentThemeOptions = themeContext!.abstractThemeOptions;
-  const currentUser = useCurrentUser();
-  const navigate = useNavigate();
-  const updateCurrentUser = useUpdateCurrentUser();
 
-  const persistUserTheme = (newThemeOptions: AbstractThemeOptions) => {
-    if (isEAForum() && currentUser) {
-      void updateCurrentUser({
-        theme: newThemeOptions as DbUser['theme'],
-      });
-    }
-  }
-  
   // When switching theme on desktop, stop event propagation so that the
   // event handler in UsersMenu doesn't close the menu, and you can try
   // multiple themes without having to reopen it.
@@ -60,42 +42,22 @@ const ThemePickerMenu = ({children}: {
 
     const newThemeOptions = {...currentThemeOptions, name};
     themeContext.setThemeOptions(newThemeOptions);
-    persistUserTheme(newThemeOptions);
-  }
-
-  const openCustomizeSidebar = () => {
-    if (designChat) {
-      designChat.setIsOpen(true);
-      return;
-    }
-    navigate('/?openCustomize=1');
   }
 
   const submenu = (
     <Paper>
       <DropdownMenu>
-        {isLWorAF() &&
-          <>
-            {isLW() && <>
-              <DropdownItem
-                title="Custom Front Page"
-                onClick={openCustomizeSidebar}
-              />
-              <DropdownDivider />
-            </>}
-            {getThemeMetadata().map((themeMetadata: ThemeMetadata) =>
-              <DropdownItem
-                key={themeMetadata.name}
-                title={themeMetadata.label}
-                onClick={(event) => setThemeName(event, themeMetadata.name)}
-                icon={() => currentThemeOptions?.name === themeMetadata.name
-                  ? <ForumIcon icon="Check" className={classes.check} />
-                  : <div className={classes.notChecked} />
-                }
-              />
-            )}
-          </>
-        }
+        {getThemeMetadata().map((themeMetadata: ThemeMetadata) =>
+          <DropdownItem
+            key={themeMetadata.name}
+            title={themeMetadata.label}
+            onClick={(event) => setThemeName(event, themeMetadata.name)}
+            icon={() => currentThemeOptions?.name === themeMetadata.name
+              ? <ForumIcon icon="Check" className={classes.check} />
+              : <div className={classes.notChecked} />
+            }
+          />
+        )}
       </DropdownMenu>
     </Paper>
   );
